+++ src/listpick/ui/build_help.py
@@ -142,14 +142,15 @@ def build_help_rows(keys_dict: dict, debug: bool = False) -> list[list[str]]:
         "file_prev":                        "Go to the previous open file.",
         "sheet_next":                       "Go to the next sheet.",
         "sheet_prev":                       "Go to the previous sheet.",
+        "toggle_right_pane":                "Toggle the right pane",
     }
     sections = {
         "Navigation:": [ "cursor_down", "cursor_up", "half_page_up", "half_page_down", "page_up", "page_down", "cursor_bottom", "cursor_top", "five_up", "five_down", "scroll_right", "scroll_left", "scroll_far_right", "scroll_far_left" ],
         "Selection:": [ "toggle_select", "select_all", "select_none", "visual_selection_toggle", "visual_deselection_toggle", "enter" ],
-        "UI:": [ "toggle_footer", "redraw_screen", "decrease_lines_per_page", "increase_lines_per_page", "increase_column_width", "decrease_column_width", "notification_toggle" ],
+        "UI:": [ "toggle_footer", "redraw_screen", "decrease_lines_per_page", "increase_lines_per_page", "increase_column_width", "decrease_column_width", "notification_toggle", "toggle_right_pane"],
         "Sort:": [ "cycle_sort_method", "cycle_sort_method_reverse", "cycle_sort_order", ] ,
         "Data manipulation:": [ "delete", "delete_column", "edit", "edit_picker", "edit_ipython", "add_column_before", "add_column_after", "add_row_before", "add_row_after"],
-        "Filter and sort:": [ "filter_input", "x", "search_input", "continue_search_forward", "continue_search_backward", ] ,
+        "Filter and sort:": [ "filter_input", "search_input", "continue_search_forward", "continue_search_backward", ] ,
         "Settings:": [ "settings_input", "settings_options" ],
         "Cancel:": [ "opts_input", "opts_select", "mode_next", "mode_prev", "pipe_input", "reset_opts", "col_select", "col_select_next", "col_select_prev", "col_hide" ],
         "Save, load, copy and paste:": [ "save", "load", "open", "copy", "paste" ],
@@ -197,15 +198,20 @@ def build_help_rows(keys_dict: dict, debug: bool = False) -> list[list[str]]:
             row = [f"    {str(keys)[1:-1]}", description]
             section_rows.append(row)
         if section_rows:
-            items.append([section_name, ""])
+            items.append([f"  {section_name}", ""])
             items += section_rows
             items.append(["",""])
+
+    if debug:
+        for operation in keys_dict:
+            if operation not in help_descriptions:
+                print(f"Note that {operation} is not in the help_descriptions")
 
     return items
 
 if __name__ == "__main__":
     items = build_help_rows(picker_keys, debug=True)
-    items = build_help_rows(notification_keys, debug=True)
+    # items = build_help_rows(notification_keys, debug=True)
 
     # from listpick.listpick_app import Picker, start_curses, close_curses
     # stdscr = start_curses()